+++ src/main/resources/article_category.sql
@@ -1,6 +1,0 @@
-INSERT INTO article_category(name) VALUES('Fruits');
-INSERT INTO article_category(name) VALUES('Vegetables');
-INSERT INTO article_category(name) VALUES('Animals');
-
-
-UPDATE article_category SET created_date = NOW();