+++ src/main/resources/article_type.sql
@@ -1,4 +1,0 @@
-INSERT INTO article_type(name) VALUES('Soft');
-INSERT INTO article_type(name) VALUES('Hard');
-
-UPDATE article_type SET created_date = NOW();