+++ src/main/resources/chart_type.sql
@@ -1,5 +1,0 @@
-INSERT INTO chart_type(name, ui_type) VALUES('BarChartModel', 'bar');
-INSERT INTO chart_type(name, ui_type) VALUES('HorizontalBarChartModel', 'bar');
-INSERT INTO chart_type(name, ui_type) VALUES('PieChartModel', 'pie');
-
-UPDATE chart_type SET created_date = NOW();